refactor: Improve modularity and maintainability of team management
Extracts data models, network operations, and UI support logic for team
management into dedicated files. This separation of concerns enhances
code organization and maintainability of `DashboardTeamsRepository`
and `DashboardTeamMembersFragment`.

Also includes minor syntax fixes for CouchDB query operators (`$elemMatch`,
`$exists`, `$regex`, `$gt`) across various repository files.
