+++ app/src/main/java/org/ole/planet/myplanet/lite/dashboard/DashboardTeamsRepository.kt
@@ -6,53 +6,25 @@
 
 package org.ole.planet.myplanet.lite.dashboard
 
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import java.io.IOException
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import okhttp3.Credentials
-import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.RequestBody.Companion.toRequestBody
-import org.json.JSONArray
-import org.json.JSONObject
 import org.ole.planet.myplanet.lite.profile.StoredCredentials
-import org.ole.planet.myplanet.lite.util.BirthDateString
 import org.ole.planet.myplanet.lite.util.DateStringAdapter
-import org.ole.planet.myplanet.lite.util.nullIfBlank
 
 class DashboardTeamsRepository(
-    private val client: OkHttpClient = OkHttpClient.Builder().build(),
-    private val moshi: Moshi = Moshi.Builder()
+    client: OkHttpClient = OkHttpClient.Builder().build(),
+    moshi: Moshi = Moshi.Builder()
         .add(DateStringAdapter())
         .addLast(KotlinJsonAdapterFactory())
         .build(),
-    dispatcher: CoroutineDispatcher = Dispatchers.IO
+    dispatcher: CoroutineDispatcher = Dispatchers.IO,
 ) {
     private val dispatcher = overrideDispatcher ?: dispatcher
-    private val membershipRequestAdapter = moshi.adapter(MembershipFindRequest::class.java)
-    private val membershipResponseAdapter = moshi.adapter(MembershipFindResponse::class.java)
-    private val teamMembershipRequestAdapter = moshi.adapter(TeamMembershipFindRequest::class.java)
-    private val multipleMemberCountRequestAdapter = moshi.adapter(MultipleMemberCountFindRequest::class.java)
-    private val multipleMemberCountResponseAdapter = moshi.adapter(MultipleMemberCountFindResponse::class.java)
-    private val teamsRequestAdapter = moshi.adapter(TeamsFindRequest::class.java)
-    private val teamsResponseAdapter = moshi.adapter(TeamsFindResponse::class.java)
-    private val availableTeamsRequestAdapter = moshi.adapter(NonMemberTeamsFindRequest::class.java)
-    private val joinRequestFindAdapter = moshi.adapter(JoinRequestFindRequest::class.java)
-    private val teamJoinRequestFindAdapter = moshi.adapter(TeamJoinRequestFindRequest::class.java)
-    private val joinRequestFindResponseAdapter = moshi.adapter(JoinRequestFindResponse::class.java)
-    private val joinTeamRequestAdapter = moshi.adapter(JoinTeamRequest::class.java)
-    private val deleteJoinRequestAdapter = moshi.adapter(DeleteDocumentRequest::class.java)
-    private val deleteMembershipAdapter = moshi.adapter(DeleteDocumentRequest::class.java)
-    private val membershipBulkDeleteAdapter = moshi.adapter(BulkMembershipDeleteRequest::class.java)
-    private val membershipBulkAddAdapter = moshi.adapter(BulkMembershipAddRequest::class.java)
-    private val usersFindResponseAdapter = moshi.adapter(UsersFindResponse::class.java)
-
+    private val operations = DashboardTeamsOperations(client, moshi, teamMemberDetailsCache)
 
     suspend fun addTeamMember(
         baseUrl: String,
@@ -63,144 +35,17 @@ class DashboardTeamsRepository(
         teamType: String = "local",
         userId: String,
         userPlanetCode: String,
-    ): Result<Unit> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (teamId.isBlank()) {
-                    throw IOException("Missing team id")
-                }
-                if (teamPlanetCode.isBlank()) {
-                    throw IOException("Missing team planet code")
-                }
-                if (userId.isBlank()) {
-                    throw IOException("Missing user id")
-                }
-                if (userPlanetCode.isBlank()) {
-                    throw IOException("Missing user planet code")
-                }
-
-                val payload = membershipBulkAddAdapter.toJson(
-                    BulkMembershipAddRequest(
-                        docs = listOf(
-                            BulkMembershipAddDoc(
-                                teamId = teamId,
-                                teamPlanetCode = teamPlanetCode,
-                                teamType = teamType.nullIfBlank() ?: "local",
-                                userId = userId,
-                                userPlanetCode = userPlanetCode,
-                                docType = "membership",
-                                isLeader = false,
-                            )
-                        )
-                    )
-                )
-                val bulkAddUrl = "$normalizedBase/db/teams/_bulk_docs"
-                val requestBuilder = Request.Builder()
-                    .url(bulkAddUrl)
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    teamMemberDetailsCache.remove(teamId)
-                    Unit
-                }
-            }
-        }
-    }
-
-    private fun buildUsersFindPayload(
-        planetCode: String,
-        parentCode: String,
-        pageSize: Int,
-        skip: Int,
-        searchTerm: String?,
-        excludedUserIds: List<String>,
-    ): String {
-        val filteredExcludedIds = excludedUserIds.filter { it.isNotBlank() }
-        val selector = JSONObject()
-            .put("planetCode", planetCode)
-            .put("parentCode", parentCode)
-
-        if (filteredExcludedIds.isNotEmpty()) {
-            val excludedArray = JSONArray()
-            filteredExcludedIds.forEach { excludedArray.put(it) }
-            selector.put("_id", JSONObject().put($$"$nin", excludedArray))
-        }
-
-        if (!searchTerm.isNullOrBlank()) {
-            val regexValue = "(?i)${searchTerm.trim()}"
-            val orArray = JSONArray()
-            listOf("name", "firstName", "middleName", "lastName").forEach { field ->
-                val regexObject = JSONObject().put($$"$regex", regexValue)
-                orArray.put(JSONObject().put(field, regexObject))
-            }
-            selector.put($$"$or", orArray)
-        }
-
-        return JSONObject()
-            .put("selector", selector)
-            .put("skip", skip)
-            .put("limit", pageSize)
-            .toString()
+    ): Result<Unit> = runInDispatcher {
+        operations.addTeamMember(baseUrl, credentials, sessionCookie, teamId, teamPlanetCode, teamType, userId, userPlanetCode)
     }
 
     suspend fun fetchMemberships(
         baseUrl: String,
         credentials: StoredCredentials?,
         sessionCookie: String?,
-        username: String
-    ): Result<List<MembershipDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (username.isBlank()) {
-                    throw IOException("Missing username")
-                }
-                val selector = MembershipSelector(
-                    userId = "org.couchdb.user:$username",
-                    teamType = "local",
-                    docType = "membership",
-                    status = StatusClause(
-                        or = listOf(
-                            StatusCondition(exists = false),
-                            StatusCondition(notEquals = "archived")
-                        )
-                    )
-                )
-                val payload = membershipRequestAdapter.toJson(MembershipFindRequest(selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    membershipResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+        username: String,
+    ): Result<List<MembershipDocument>> = runInDispatcher {
+        operations.fetchMemberships(baseUrl, credentials, sessionCookie, username)
     }
 
     suspend fun fetchTeamMemberDetails(
@@ -208,52 +53,8 @@ class DashboardTeamsRepository(
         credentials: StoredCredentials?,
         sessionCookie: String?,
         teamId: String,
-    ): Result<List<TeamMemberDetails>> {
-        return withContext(dispatcher) {
-            runCatching {
-                teamMemberDetailsCache[teamId]?.let { return@runCatching it }
-
-                val memberships = fetchTeamMembers(baseUrl, credentials, sessionCookie, teamId)
-                    .getOrThrow()
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-
-                val validMemberships = memberships.filter { !it.userId.isNullOrBlank() }
-                val userIds = validMemberships.mapNotNull { it.userId }
-
-                val profiles = if (userIds.isNotEmpty()) {
-                    fetchUserProfiles(normalizedBase, credentials, sessionCookie, userIds).getOrNull() ?: emptyList()
-                } else {
-                    emptyList()
-                }
-
-                val profileMap = profiles.associateBy { it._id }
-
-                val details = validMemberships.mapNotNull { member ->
-                    val userId = member.userId ?: return@mapNotNull null
-                    val username = userId.substringAfter("org.couchdb.user:")
-                    if (username.isBlank()) {
-                        return@mapNotNull null
-                    }
-                    val profile = profileMap[userId]
-                    val fullName = listOfNotNull(
-                        profile?.firstName,
-                        profile?.middleName,
-                        profile?.lastName,
-                    ).joinToString(" ").ifBlank { username }
-
-                    TeamMemberDetails(
-                        username = username,
-                        fullName = fullName,
-                        isLeader = member.isLeader == true,
-                        hasAvatar = profile?.attachments?.image != null,
-                        membership = member,
-                    )
-                }
-
-                teamMemberDetailsCache[teamId] = details
-                details
-            }
-        }
+    ): Result<List<TeamMemberDetails>> = runInDispatcher {
+        operations.fetchTeamMemberDetails(baseUrl, credentials, sessionCookie, teamId)
     }
 
     suspend fun fetchTeamMembers(
@@ -261,47 +62,8 @@ class DashboardTeamsRepository(
         credentials: StoredCredentials?,
         sessionCookie: String?,
         teamId: String,
-    ): Result<List<MembershipDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (teamId.isBlank()) {
-                    throw IOException("Missing team id")
-                }
-
-                val selector = TeamMembershipSelector(
-                    teamId = teamId,
-                    docType = "membership",
-                    status = StatusClause(
-                        or = listOf(
-                            StatusCondition(exists = false),
-                            StatusCondition(notEquals = "archived"),
-                        ),
-                    ),
-                )
-                val payload = teamMembershipRequestAdapter.toJson(TeamMembershipFindRequest(selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    membershipResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+    ): Result<List<MembershipDocument>> = runInDispatcher {
+        operations.fetchTeamMembers(baseUrl, credentials, sessionCookie, teamId)
     }
 
     suspend fun removeTeamMember(
@@ -309,106 +71,17 @@ class DashboardTeamsRepository(
         credentials: StoredCredentials?,
         sessionCookie: String?,
         membership: MembershipDocument,
-    ): Result<Unit> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-
-                val id = membership.id.nullIfBlank() ?: throw IOException("Missing membership id")
-                val revision = membership.revision.nullIfBlank()
-                    ?: throw IOException("Missing membership revision")
-                val teamId = membership.teamId.nullIfBlank() ?: throw IOException("Missing team id")
-                val teamPlanetCode = membership.teamPlanetCode.nullIfBlank()
-                    ?: throw IOException("Missing team planet code")
-                val userId = membership.userId.nullIfBlank() ?: throw IOException("Missing user id")
-                val userPlanetCode = membership.userPlanetCode.nullIfBlank()
-                    ?: throw IOException("Missing user planet code")
-                val teamType = membership.teamType.nullIfBlank() ?: "local"
-                val docType = membership.docType.nullIfBlank() ?: "membership"
-
-                val payload = membershipBulkDeleteAdapter.toJson(
-                    BulkMembershipDeleteRequest(
-                        docs = listOf(
-                            BulkMembershipDeleteDoc(
-                                id = id,
-                                revision = revision,
-                                teamId = teamId,
-                                teamPlanetCode = teamPlanetCode,
-                                teamType = teamType,
-                                userId = userId,
-                                userPlanetCode = userPlanetCode,
-                                docType = docType,
-                                isLeader = false,
-                                deleted = true,
-                            ),
-                        ),
-                    ),
-                )
-                val bulkDeleteUrl = "$normalizedBase/db/teams/_bulk_docs"
-                val requestBuilder = Request.Builder()
-                    .url(bulkDeleteUrl)
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    teamMemberDetailsCache.remove(teamId)
-                    Unit
-                }
-            }
-        }
+    ): Result<Unit> = runInDispatcher {
+        operations.removeTeamMember(baseUrl, credentials, sessionCookie, membership)
     }
 
     suspend fun fetchTeams(
         baseUrl: String,
         credentials: StoredCredentials?,
         sessionCookie: String?,
-        teamIds: List<String>
-    ): Result<List<TeamDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (teamIds.isEmpty()) {
-                    return@runCatching emptyList()
-                }
-                val selector = TeamsSelector(
-                    status = "active",
-                    type = "team",
-                    teamType = "local",
-                    ids = IdsInClause(ids = teamIds)
-                )
-                val payload = teamsRequestAdapter.toJson(TeamsFindRequest(selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    teamsResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+        teamIds: List<String>,
+    ): Result<List<TeamDocument>> = runInDispatcher {
+        operations.fetchTeams(baseUrl, credentials, sessionCookie, teamIds)
     }
 
     suspend fun fetchAvailableTeams(
@@ -417,46 +90,9 @@ class DashboardTeamsRepository(
         sessionCookie: String?,
         excludedTeamIds: List<String>,
         skip: Int = 0,
-        limit: Int = 25
-    ): Result<List<TeamDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-
-                val selector = NonMemberTeamsSelector(
-                    ids = IdsNotInClause(ids = excludedTeamIds),
-                    status = "active",
-                    type = "team",
-                    teamType = "local"
-                )
-                val payload = availableTeamsRequestAdapter.toJson(
-                    NonMemberTeamsFindRequest(
-                        selector = selector,
-                        limit = limit,
-                        skip = skip,
-                    )
-                )
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    teamsResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+        limit: Int = 25,
+    ): Result<List<TeamDocument>> = runInDispatcher {
+        operations.fetchAvailableTeams(baseUrl, credentials, sessionCookie, excludedTeamIds, skip, limit)
     }
 
     suspend fun fetchJoinRequests(
@@ -464,42 +100,8 @@ class DashboardTeamsRepository(
         credentials: StoredCredentials?,
         sessionCookie: String?,
         userId: String,
-    ): Result<List<JoinRequestDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (userId.isBlank()) {
-                    throw IOException("Missing user id")
-                }
-
-                val selector = JoinRequestSelector(
-                    docType = "request",
-                    teamType = "local",
-                    userId = userId,
-                )
-                val payload = joinRequestFindAdapter.toJson(JoinRequestFindRequest(selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    joinRequestFindResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+    ): Result<List<JoinRequestDocument>> = runInDispatcher {
+        operations.fetchJoinRequests(baseUrl, credentials, sessionCookie, userId)
     }
 
     suspend fun fetchTeamJoinRequests(
@@ -508,51 +110,8 @@ class DashboardTeamsRepository(
         sessionCookie: String?,
         teamId: String,
         teamPlanetCode: String,
-    ): Result<List<JoinRequestDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (teamId.isBlank()) {
-                    throw IOException("Missing team id")
-                }
-                if (teamPlanetCode.isBlank()) {
-                    throw IOException("Missing team planet code")
-                }
-
-                val selector = TeamJoinRequestSelector(
-                    teamId = teamId,
-                    teamPlanetCode = teamPlanetCode,
-                    docType = "request",
-                    status = StatusClause(
-                        or = listOf(
-                            StatusCondition(exists = false),
-                            StatusCondition(notEquals = "archived")
-                        )
-                    )
-                )
-                val payload = teamJoinRequestFindAdapter.toJson(TeamJoinRequestFindRequest(selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    joinRequestFindResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+    ): Result<List<JoinRequestDocument>> = runInDispatcher {
+        operations.fetchTeamJoinRequests(baseUrl, credentials, sessionCookie, teamId, teamPlanetCode)
     }
 
     suspend fun hasExistingJoinRequest(
@@ -561,104 +120,17 @@ class DashboardTeamsRepository(
         sessionCookie: String?,
         teamId: String,
         userId: String,
-    ): Result<Boolean> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (teamId.isBlank()) {
-                    throw IOException("Missing team id")
-                }
-                if (userId.isBlank()) {
-                    throw IOException("Missing user id")
-                }
-
-                val selector = JoinRequestSelector(
-                    docType = "request",
-                    teamType = "local",
-                    userId = userId,
-                    teamId = teamId,
-                )
-                val payload = joinRequestFindAdapter.toJson(JoinRequestFindRequest(selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    val docs = joinRequestFindResponseAdapter.fromJson(body)?.docs
-                    !docs.isNullOrEmpty()
-                }
-            }
-        }
+    ): Result<Boolean> = runInDispatcher {
+        operations.hasExistingJoinRequest(baseUrl, credentials, sessionCookie, teamId, userId)
     }
 
     suspend fun fetchMemberCounts(
         baseUrl: String,
         credentials: StoredCredentials?,
         sessionCookie: String?,
-        teamIds: List<String>
-    ): Result<Map<String, Int>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-
-                if (teamIds.isEmpty()) return@runCatching emptyMap()
-
-                val selector = MultipleMemberCountSelector(
-                    teamId = IdsInClause(ids = teamIds),
-                    docType = "membership",
-                    status = StatusClause(
-                        or = listOf(
-                            StatusCondition(exists = false),
-                            StatusCondition(notEquals = "archived")
-                        )
-                    )
-                )
-                val payload = multipleMemberCountRequestAdapter.toJson(
-                    MultipleMemberCountFindRequest(selector = selector, fields = listOf("_id", "teamId"), limit = 50000)
-                )
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                val allDocs = client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    multipleMemberCountResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-
-                val results = mutableMapOf<String, Int>()
-                allDocs.forEach { doc ->
-                    doc.teamId?.let { tId ->
-                        results[tId] = (results[tId] ?: 0) + 1
-                    }
-                }
-                results
-            }
-        }
+        teamIds: List<String>,
+    ): Result<Map<String, Int>> = runInDispatcher {
+        operations.fetchMemberCounts(baseUrl, credentials, sessionCookie, teamIds)
     }
 
     suspend fun requestTeamMembership(
@@ -666,86 +138,8 @@ class DashboardTeamsRepository(
         credentials: StoredCredentials?,
         sessionCookie: String?,
         request: JoinTeamRequest,
-    ): Result<Unit> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (request.teamId.isBlank()) {
-                    throw IOException("Missing team id")
-                }
-                if (request.userId.isBlank()) {
-                    throw IOException("Missing user id")
-                }
-
-                val payload = joinTeamRequestAdapter.toJson(request)
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    teamMemberDetailsCache.remove(request.teamId)
-                    Unit
-                }
-            }
-        }
-    }
-
-    private fun fetchUserProfile(
-        baseUrl: String,
-        username: String,
-        credentials: StoredCredentials?,
-        sessionCookie: String?,
-    ): TeamMemberProfileDetails? {
-        if (baseUrl.isBlank() || username.isBlank()) {
-            return null
-        }
-        val requestBuilder = Request.Builder()
-            .url("$baseUrl/db/_users/org.couchdb.user:$username")
-            .get()
-        credentials?.let {
-            requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-        }
-        sessionCookie.nullIfBlank()?.let { cookie ->
-            requestBuilder.addHeader("Cookie", cookie)
-        }
-
-        return try {
-            client.newCall(requestBuilder.build()).execute().use { response ->
-                if (!response.isSuccessful) {
-                    return null
-                }
-                val body = response.body.string().nullIfBlank() ?: return null
-                val json = JSONObject(body)
-                val attachments = json.optJSONObject("_attachments")
-                TeamMemberProfileDetails(
-                    username = json.optString("name").nullIfBlank() ?: username,
-                    firstName = json.optString("firstName").nullIfBlank(),
-                    middleName = json.optString("middleName").nullIfBlank(),
-                    lastName = json.optString("lastName").nullIfBlank(),
-                    email = json.optString("email").nullIfBlank(),
-                    phoneNumber = json.optString("phoneNumber").nullIfBlank(),
-                    language = json.optString("language").nullIfBlank(),
-                    level = json.optString("level").nullIfBlank(),
-                    gender = json.optString("gender").nullIfBlank(),
-                    birthDate = json.optString("birthDate").nullIfBlank(),
-                    hasAvatar = attachments?.optJSONObject("img") != null,
-                )
-            }
-        } catch (_: IOException) {
-            null
-        }
+    ): Result<Unit> = runInDispatcher {
+        operations.requestTeamMembership(baseUrl, credentials, sessionCookie, request)
     }
 
     suspend fun fetchTeamMemberProfileDetails(
@@ -753,17 +147,8 @@ class DashboardTeamsRepository(
         credentials: StoredCredentials?,
         sessionCookie: String?,
         username: String,
-    ): Result<TeamMemberProfileDetails> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing base url")
-                }
-                fetchUserProfile(normalizedBase, username, credentials, sessionCookie)
-                    ?: throw IOException("Profile not found")
-            }
-        }
+    ): Result<TeamMemberProfileDetails> = runInDispatcher {
+        operations.fetchTeamMemberProfileDetails(baseUrl, credentials, sessionCookie, username)
     }
 
     suspend fun cancelJoinRequest(
@@ -772,41 +157,8 @@ class DashboardTeamsRepository(
         sessionCookie: String?,
         documentId: String,
         revision: String,
-    ): Result<Unit> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (documentId.isBlank()) {
-                    throw IOException("Missing document id")
-                }
-                if (revision.isBlank()) {
-                    throw IOException("Missing document revision")
-                }
-
-                val payload = deleteJoinRequestAdapter.toJson(
-                    DeleteDocumentRequest(id = documentId, revision = revision, deleted = true)
-                )
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    teamMemberDetailsCache.clear()
-                }
-            }
-        }
+    ): Result<Unit> = runInDispatcher {
+        operations.cancelJoinRequest(baseUrl, credentials, sessionCookie, documentId, revision)
     }
 
     suspend fun cancelMembership(
@@ -815,376 +167,17 @@ class DashboardTeamsRepository(
         sessionCookie: String?,
         documentId: String,
         revision: String,
-    ): Result<Unit> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                if (documentId.isBlank()) {
-                    throw IOException("Missing document id")
-                }
-                if (revision.isBlank()) {
-                    throw IOException("Missing document revision")
-                }
-
-                val payload = deleteMembershipAdapter.toJson(
-                    DeleteDocumentRequest(id = documentId, revision = revision, deleted = true)
-                )
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/teams")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                credentials?.let {
-                    requestBuilder.addHeader("Authorization", Credentials.basic(it.username, it.password))
-                }
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    teamMemberDetailsCache.clear()
-                }
-            }
-        }
+    ): Result<Unit> = runInDispatcher {
+        operations.cancelMembership(baseUrl, credentials, sessionCookie, documentId, revision)
     }
-
-    @JsonClass(generateAdapter = true)
-    data class MembershipFindRequest(val selector: MembershipSelector)
-
-    @JsonClass(generateAdapter = true)
-    data class MembershipSelector(
-        val userId: String,
-        val teamType: String,
-        val docType: String,
-        val status: StatusClause
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class TeamMembershipFindRequest(val selector: TeamMembershipSelector)
-
-    @JsonClass(generateAdapter = true)
-    data class TeamMembershipSelector(
-        val teamId: String,
-        val docType: String,
-        val status: StatusClause,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class StatusClause(
-        @param:Json(name = $$"$or") val or: List<StatusCondition>
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class StatusCondition(
-        @param:Json(name = $$"$exists") val exists: Boolean? = null,
-        @param:Json(name = $$"$ne") val notEquals: String? = null
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class MembershipFindResponse(val docs: List<MembershipDocument>?)
-
-    @JsonClass(generateAdapter = true)
-    data class MultipleMemberCountFindRequest(
-        val selector: MultipleMemberCountSelector,
-        val fields: List<String>,
-        val limit: Int? = null
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class MultipleMemberCountSelector(
-        val teamId: IdsInClause,
-        val docType: String,
-        val status: StatusClause
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class MemberTeamIdDocument(
-        @param:Json(name = "_id") val id: String?,
-        val teamId: String?
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class MultipleMemberCountFindResponse(val docs: List<MemberTeamIdDocument>?)
-
-    @JsonClass(generateAdapter = true)
-    data class MembershipDocument(
-        @param:Json(name = "_id") val id: String?,
-        @param:Json(name = "_rev") val revision: String?,
-        val teamId: String?,
-        val userId: String?,
-        val teamPlanetCode: String?,
-        val teamType: String?,
-        val userPlanetCode: String?,
-        val docType: String?,
-        val isLeader: Boolean?,
-        val status: String?
-    )
-
-    data class TeamMemberDetails(
-        val username: String?,
-        val fullName: String?,
-        val isLeader: Boolean,
-        val hasAvatar: Boolean,
-        val membership: MembershipDocument?,
-    )
-
-    data class TeamMemberProfileDetails(
-        val username: String,
-        val firstName: String?,
-        val middleName: String?,
-        val lastName: String?,
-        val email: String?,
-        val phoneNumber: String?,
-        val language: String?,
-        val level: String?,
-        val gender: String?,
-        val birthDate: String?,
-        val hasAvatar: Boolean,
-    ) {
-        val fullName: String?
-            get() {
-                val parts = listOfNotNull(firstName, middleName, lastName)
-                    .filter { it.isNotBlank() }
-                return if (parts.isEmpty()) null else parts.joinToString(" ")
-            }
-    }
-
-    @JsonClass(generateAdapter = true)
-    data class TeamsFindRequest(val selector: TeamsSelector)
-
-    @JsonClass(generateAdapter = true)
-    data class TeamsSelector(
-        val status: String,
-        val type: String,
-        val teamType: String,
-        @param:Json(name = "_id") val ids: IdsInClause
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class IdsInClause(
-        @param:Json(name = $$"$in") val ids: List<String>
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class NonMemberTeamsFindRequest(
-        val selector: NonMemberTeamsSelector,
-        val limit: Int? = null,
-        val skip: Int? = null,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class NonMemberTeamsSelector(
-        @param:Json(name = "_id") val ids: IdsNotInClause?,
-        val status: String,
-        val type: String,
-        val teamType: String? = null
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class IdsNotInClause(
-        @param:Json(name = $$"$nin") val ids: List<String>
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class TeamsFindResponse(val docs: List<TeamDocument>?)
-
-    @JsonClass(generateAdapter = true)
-    data class TeamDocument(
-        @param:Json(name = "_id") val id: String?,
-        @param:Json(name = "_rev") val revision: String?,
-        val limit: Int?,
-        val status: String?,
-        val type: String?,
-        val teamType: String?,
-        val name: String?,
-        @param:Json(name = "teamName") val teamName: String?,
-        @param:Json(name = "planetCode") val planetCode: String?,
-        val teamPlanetCode: String?,
-        val description: String?,
-        val services: String?,
-        val rules: String?,
-        val requests: List<Any>?,
-        val createdDate: Long?,
-        val createdBy: String?,
-        @param:Json(name = "parentCode") val parentCode: String?,
-        @param:Json(name = "public") val isPublic: Boolean?,
-        @param:Json(name = "memberCount") val memberCount: Int?,
-        @param:Json(name = "membersCount") val membersCount: Int?,
-        val members: List<Any>?
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class JoinRequestFindRequest(val selector: JoinRequestSelector)
-
-    @JsonClass(generateAdapter = true)
-    data class JoinRequestSelector(
-        val docType: String,
-        val teamType: String? = null,
-        val teamId: String? = null,
-        val userId: String,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class TeamJoinRequestFindRequest(val selector: TeamJoinRequestSelector)
-
-    @JsonClass(generateAdapter = true)
-    data class TeamJoinRequestSelector(
-        val teamId: String,
-        val teamPlanetCode: String,
-        val docType: String,
-        val status: StatusClause,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class JoinRequestDocument(
-        @param:Json(name = "_id") val id: String?,
-        @param:Json(name = "_rev") val revision: String?,
-        val docType: String?,
-        val teamId: String?,
-        val teamType: String?,
-        val teamPlanetCode: String?,
-        val userId: String?,
-        val userPlanetCode: String?,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class JoinRequestFindResponse(val docs: List<JoinRequestDocument>?)
-
-    @JsonClass(generateAdapter = true)
-    data class JoinTeamRequest(
-        val docType: String = "request",
-        val teamId: String,
-        val teamType: String = "local",
-        val teamPlanetCode: String?,
-        val userId: String,
-        val userPlanetCode: String?,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class DeleteDocumentRequest(
-        @param:Json(name = "_id") val id: String,
-        @param:Json(name = "_rev") val revision: String,
-        @param:Json(name = "_deleted") val deleted: Boolean,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class BulkMembershipDeleteRequest(val docs: List<BulkMembershipDeleteDoc>)
-
-    @JsonClass(generateAdapter = true)
-    data class BulkMembershipDeleteDoc(
-        @param:Json(name = "_id") val id: String,
-        @param:Json(name = "_rev") val revision: String,
-        val teamId: String,
-        val teamPlanetCode: String,
-        val teamType: String,
-        val userId: String,
-        val userPlanetCode: String,
-        val docType: String,
-        val isLeader: Boolean,
-        @param:Json(name = "_deleted") val deleted: Boolean,
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class BulkMembershipAddRequest(val docs: List<BulkMembershipAddDoc>)
-
-    @JsonClass(generateAdapter = true)
-    data class BulkMembershipAddDoc(
-        val teamId: String,
-        val teamPlanetCode: String,
-        val teamType: String,
-        val userId: String,
-        val userPlanetCode: String,
-        val docType: String,
-        val isLeader: Boolean,
-    )
-
-    companion object {
-        private val JSON_MEDIA_TYPE = "application/json; charset=utf-8".toMediaType()
-        private val teamMemberDetailsCache = java.util.concurrent.ConcurrentHashMap<String, List<TeamMemberDetails>>()
-
-        @androidx.annotation.VisibleForTesting
-        var overrideDispatcher: CoroutineDispatcher? = null
-
-        @androidx.annotation.VisibleForTesting
-        fun resetCacheForTesting() {
-            teamMemberDetailsCache.clear()
-            overrideDispatcher = null
-        }
-    }
-
-    @JsonClass(generateAdapter = true)
-    data class UserDocument(
-        @param:Json(name = "_id") val _id: String?,
-        @param:Json(name = "_attachments") val attachments: Attachments?,
-        @param:Json(name = "planetCode") val planetCode: String?,
-        @param:Json(name = "parentCode") val parentCode: String?,
-        val firstName: String?,
-        val middleName: String?,
-        val lastName: String?,
-        val email: String?,
-        val language: String?,
-        val phoneNumber: String?,
-        @param:BirthDateString val birthDate: Long?,
-        val gender: String?,
-        val level: String?
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class Attachments(
-        @param:Json(name = "img") val image: Attachment?
-    )
-
-    @JsonClass(generateAdapter = true)
-    data class Attachment(
-        @param:Json(name = "content_type") val contentType: String?,
-        val revpos: Int?,
-        val digest: String?,
-        val length: Long?,
-        val stub: Boolean?
-    )
 
     suspend fun fetchUserProfiles(
         baseUrl: String,
         credentials: StoredCredentials?,
         sessionCookie: String?,
-        userIds: List<String>
-    ): Result<List<UserDocument>> {
-        return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                val basicAuth = credentials?.let { Credentials.basic(it.username, it.password) }
-                    ?: throw IOException("Missing credentials for basic auth")
-                if (userIds.isEmpty()) {
-                    return@runCatching emptyList()
-                }
-
-                val selector = UserIdSelector(ids = IdsInClause(ids = userIds))
-                val payload = moshi.adapter(UsersFindRequest::class.java).toJson(UsersFindRequest(selector = selector))
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/_users/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                    .header("Authorization", basicAuth)
-                    .header("Content-Type", "application/json")
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    moshi.adapter(UsersFindResponse::class.java).fromJson(body)?.docs ?: emptyList()
-                }
-            }
-        }
+        userIds: List<String>,
+    ): Result<List<UserDocument>> = runInDispatcher {
+        operations.fetchUserProfiles(baseUrl, credentials, sessionCookie, userIds)
     }
 
     suspend fun fetchAllUsers(
@@ -1197,59 +190,36 @@ class DashboardTeamsRepository(
         skip: Int = 0,
         searchTerm: String? = null,
         excludedUserIds: List<String> = emptyList(),
-    ): Result<List<UserDocument>> {
+    ): Result<List<UserDocument>> = runInDispatcher {
+        operations.fetchAllUsers(
+            baseUrl = baseUrl,
+            credentials = credentials,
+            sessionCookie = sessionCookie,
+            planetCode = planetCode,
+            parentCode = parentCode,
+            pageSize = pageSize,
+            skip = skip,
+            searchTerm = searchTerm,
+            excludedUserIds = excludedUserIds,
+        )
+    }
+
+    private suspend fun <T> runInDispatcher(block: () -> T): Result<T> {
         return withContext(dispatcher) {
-            runCatching {
-                val normalizedBase = baseUrl.trim().trimEnd('/')
-                if (normalizedBase.isEmpty()) {
-                    throw IOException("Missing server base URL")
-                }
-                val basicAuth = credentials?.let { Credentials.basic(it.username, it.password) }
-                    ?: throw IOException("Missing credentials for basic auth")
-                if (pageSize <= 0) {
-                    return@runCatching emptyList()
-                }
-
-                val filteredPlanet = planetCode.nullIfBlank()
-                    ?: throw IOException("Missing planet code for user search")
-                val filteredParent = parentCode.nullIfBlank()
-                    ?: throw IOException("Missing parent code for user search")
-
-                val payload = buildUsersFindPayload(
-                    planetCode = filteredPlanet,
-                    parentCode = filteredParent,
-                    pageSize = pageSize,
-                    skip = skip,
-                    searchTerm = searchTerm,
-                    excludedUserIds = excludedUserIds,
-                )
-                val requestBuilder = Request.Builder()
-                    .url("$normalizedBase/db/_users/_find")
-                    .post(payload.toRequestBody(JSON_MEDIA_TYPE))
-                    .header("Authorization", basicAuth)
-                    .header("Content-Type", "application/json")
-                sessionCookie.nullIfBlank()?.let { cookie ->
-                    requestBuilder.addHeader("Cookie", cookie)
-                }
-
-                client.newCall(requestBuilder.build()).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw IOException("Unexpected response ${response.code}")
-                    }
-                    val body = response.body.string()
-                    usersFindResponseAdapter.fromJson(body)?.docs ?: emptyList()
-                }
-            }
+            runCatching(block)
         }
     }
 
-    @JsonClass(generateAdapter = true)
-    data class UsersFindRequest(val selector: UserIdSelector)
+    companion object {
+        private val teamMemberDetailsCache = java.util.concurrent.ConcurrentHashMap<String, List<TeamMemberDetails>>()
 
-    @JsonClass(generateAdapter = true)
-    data class UserIdSelector(@param:Json(name = "_id") val ids: IdsInClause)
+        @androidx.annotation.VisibleForTesting
+        var overrideDispatcher: CoroutineDispatcher? = null
 
-    @JsonClass(generateAdapter = true)
-    data class UsersFindResponse(val docs: List<UserDocument>?)
-
+        @androidx.annotation.VisibleForTesting
+        fun resetCacheForTesting() {
+            teamMemberDetailsCache.clear()
+            overrideDispatcher = null
+        }
+    }
 }
