+++ app/src/test/java/org/ole/planet/myplanet/lite/dashboard/DashboardResourcesRepositoryTest.kt
@@ -46,6 +46,6 @@ class DashboardResourcesRepositoryTest {
         val selector = body.getJSONObject("selector")
         assertEquals(false, selector.has("createdDate"))
         assertEquals(true, selector.has("_id"))
-        assertEquals(true, selector.getJSONObject("_id").has("\$gt"))
+        assertEquals(true, selector.getJSONObject("_id").has($$"$gt"))
     }
 }
